+++ src/public/options.ts
@@ -483,7 +483,7 @@ export type UndoStateChangeListener = (
  * In general instead of using this interface you should be listening to the
  * corresponding event on `MathfieldElement`, i.e.
  * ```javascript
-mfe.addEventListener('change', (ev) => {
+mfe.addEventListener('input', (ev) => {
     console.log(ev.target.value);
 });
  * ```
@@ -500,6 +500,7 @@ export interface MathfieldListeners {
     onSelectionDidChange: (sender: Mathfield) => void;
     onUndoStateWillChange: UndoStateChangeListener;
     onUndoStateDidChange: UndoStateChangeListener;
+    onCommit: (sender: Mathfield) => void;
     onModeChange: (sender: Mathfield, mode: ParseMode) => void;
     onVirtualKeyboardToggle: (
         sender: Mathfield,